Ending main microservices.
