--- menu-microservice-category/Dockerfile
@@ -1,4 +1,0 @@
-FROM openjdk:8
-ADD target/menu-microservice-category-0.0.1-SNAPSHOT.jar menu-microservice-category-0.0.1-SNAPSHOT.jar
-ENTRYPOINT ["java", "-jar", "menu-microservice-category-0.0.1-SNAPSHOT.jar"]
-EXPOSE 8080
