--- menu-microservice-category/src/main/java/br/com/menu/menu/microservice/category/MenuMicroserviceCategoryApplication.java
@@ -1,12 +1,0 @@
-package br.com.menu.menu.microservice.category;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MenuMicroserviceCategoryApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(MenuMicroserviceCategoryApplication.class, args);
-	}
-}
