--- menu-microservice-category/src/main/java/br/com/menu/menu/microservice/category/dto/CategoryDto.java
@@ -1,28 +1,0 @@
-package br.com.menu.menu.microservice.category.dto;
-
-import java.io.Serializable;
-
-public class CategoryDto implements Serializable {
-
-    private static final long serialVersionUID = 1L;
-
-    private String name;
-    
-    private Long categoryId;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public Long getCategoryId() {
-        return categoryId;
-    }
-
-    public void setCategoryId(Long categoryId) {
-        this.categoryId = categoryId;
-    }
-}
