--- menu-microservice-category/src/main/resources/application.properties
@@ -1,8 +1,0 @@
-#spring.datasource.url=jdbc:mysql://mysql:3306/test
-#spring.datasource.username=root
-#spring.datasource.password=hellokoding
-#
-#spring.jpa.hibernate.ddl-auto=none
-#spring.jpa.database-platform=org.hibernate.dialect.MySQL5Dialect
-#spring.jpa.generate-ddl=true
-#spring.jpa.show-sql=true
