--- menu-microservice-product-type/src/main/java/br/com/menu/menu/microservice/product/type/MenuMicrosericeProductTypeApplication.java
@@ -1,12 +1,0 @@
-package br.com.menu.menu.microservice.product.type;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MenuMicrosericeProductTypeApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(MenuMicrosericeProductTypeApplication.class, args);
-	}
-}
