--- menu-microservice-product-type/src/main/java/br/com/menu/menu/microservice/product/type/enums/Category.java
@@ -1,7 +1,0 @@
-package br.com.menu.menu.microservice.product.type.enums;
-
-public enum Category {
-
-    bar,
-    kitchen
-}
