+++ menu-microservice-product-type/src/main/java/br/com/menu/microservice/product/type/MenuMicrosericeProductTypeApplication.java
@@ -1,0 +1,12 @@
+package br.com.menu.microservice.product.type;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class MenuMicrosericeProductTypeApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(MenuMicrosericeProductTypeApplication.class, args);
+	}
+}
