+++ menu-microservice-product-type/src/main/java/br/com/menu/microservice/product/type/enums/Category.java
@@ -1,0 +1,7 @@
+package br.com.menu.microservice.product.type.enums;
+
+public enum Category {
+
+    bar,
+    kitchen
+}
