+++ menu-microservice-product-type/src/main/java/br/com/menu/microservice/product/type/repository/ProductTypeRepository.java
@@ -1,0 +1,9 @@
+package br.com.menu.microservice.product.type.repository;
+
+import br.com.menu.microservice.product.type.model.ProductTypeEntity;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ProductTypeRepository extends CrudRepository<ProductTypeEntity, Long> {
+}
