+++ menu-microservice-product-type/src/main/resources/application.properties
@@ -2,7 +2,7 @@ spring.datasource.url=jdbc:mysql://mysql:3306/test
 spring.datasource.username=root
 spring.datasource.password=hellokoding
 
-spring.jpa.hibernate.ddl-auto=none
+spring.jpa.hibernate.ddl-auto=create-drop
 spring.jpa.database-platform=org.hibernate.dialect.MySQL5Dialect
 spring.jpa.generate-ddl=true
 spring.jpa.show-sql=true
