--- menu-microservice-product/src/main/java/br/com/menu/menu/microservice/product/MenuMicroserviceProductApplication.java
@@ -1,12 +1,0 @@
-package br.com.menu.menu.microservice.product;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MenuMicroserviceProductApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(MenuMicroserviceProductApplication.class, args);
-	}
-}
