--- menu-microservice-product/src/main/java/br/com/menu/menu/microservice/product/dto/ProductDto.java
@@ -1,55 +1,0 @@
-package br.com.menu.menu.microservice.product.dto;
-
-import javax.persistence.Column;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import java.io.Serializable;
-import java.math.BigDecimal;
-
-public class ProductDto implements Serializable {
-
-    private static final long serialVersionUID = 1L;
-
-    private Long productId;
-
-    // TODO adicionar Foreign Key ProductType
-
-    private String description;
-
-    private String name;
-
-    private BigDecimal price;
-
-    public Long getProductId() {
-        return productId;
-    }
-
-    public void setProductId(Long productId) {
-        this.productId = productId;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public BigDecimal getPrice() {
-        return price;
-    }
-
-    public void setPrice(BigDecimal price) {
-        this.price = price;
-    }
-}
