--- menu-microservice-product/src/main/java/br/com/menu/menu/microservice/product/repository/ProductRepository.java
@@ -1,10 +1,0 @@
-package br.com.menu.menu.microservice.product.repository;
-
-import br.com.menu.menu.microservice.product.model.ProductEntity;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ProductRepository extends CrudRepository<ProductEntity, Long> {
-
-}
