+++ menu-microservice-product/src/main/java/br/com/menu/microservice/product/MenuMicroserviceProductApplication.java
@@ -1,0 +1,20 @@
+package br.com.menu.microservice.product;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.web.client.RestTemplateBuilder;
+import org.springframework.context.annotation.Bean;
+import org.springframework.web.client.RestTemplate;
+
+@SpringBootApplication
+public class MenuMicroserviceProductApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(MenuMicroserviceProductApplication.class, args);
+	}
+
+	@Bean
+	public RestTemplate restTemplate(RestTemplateBuilder builder) {
+		return builder.build();
+	}
+}
