+++ menu-microservice-product/src/main/java/br/com/menu/microservice/product/dto/ProductRequestDto.java
@@ -1,0 +1,63 @@
+package br.com.menu.microservice.product.dto;
+
+import javax.persistence.Column;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import java.io.Serializable;
+import java.math.BigDecimal;
+
+public class ProductRequestDto implements Serializable {
+
+    private static final long serialVersionUID = 1L;
+
+    private Long productId;
+
+    private Long productTypeId;
+
+    private String description;
+
+    private String name;
+
+    private BigDecimal price;
+
+    public Long getProductId() {
+        return productId;
+    }
+
+    public void setProductId(Long productId) {
+        this.productId = productId;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public BigDecimal getPrice() {
+        return price;
+    }
+
+    public void setPrice(BigDecimal price) {
+        this.price = price;
+    }
+
+    public Long getProductTypeId() {
+        return productTypeId;
+    }
+
+    public void setProductTypeId(Long productTypeId) {
+        this.productTypeId = productTypeId;
+    }
+}
