+++ menu-microservice-product/src/main/java/br/com/menu/microservice/product/dto/ProductResponseDto.java
@@ -1,0 +1,73 @@
+package br.com.menu.microservice.product.dto;
+
+import br.com.menu.microservice.product.enums.Category;
+import com.fasterxml.jackson.annotation.JsonInclude;
+
+import java.io.Serializable;
+import java.math.BigDecimal;
+
+@JsonInclude(JsonInclude.Include.NON_NULL)
+public class ProductResponseDto implements Serializable {
+
+    private static final long serialVersionUID = 1L;
+
+    private Long productId;
+
+    private String description;
+
+    private String name;
+
+    private BigDecimal price;
+
+    private Category category;
+
+    private String productType;
+
+    public Long getProductId() {
+        return productId;
+    }
+
+    public void setProductId(Long productId) {
+        this.productId = productId;
+    }
+
+    public String getProductType() {
+        return productType;
+    }
+
+    public void setProductType(String productType) {
+        this.productType = productType;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public BigDecimal getPrice() {
+        return price;
+    }
+
+    public void setPrice(BigDecimal price) {
+        this.price = price;
+    }
+
+    public Category getCategory() {
+        return category;
+    }
+
+    public void setCategory(Category category) {
+        this.category = category;
+    }
+}
