+++ menu-microservice-product/src/main/java/br/com/menu/microservice/product/enums/Category.java
@@ -1,0 +1,7 @@
+package br.com.menu.microservice.product.enums;
+
+public enum Category {
+
+    bar,
+    kitchen
+}
