+++ menu-microservice-product/src/main/java/br/com/menu/microservice/product/mapper/ProductTypeMapper.java
@@ -1,0 +1,14 @@
+package br.com.menu.microservice.product.mapper;
+
+import br.com.menu.microservice.product.dto.ProductTypeDto;
+import br.com.menu.microservice.product.model.ProductTypeEntity;
+import org.mapstruct.Mapper;
+
+@Mapper(componentModel = "spring")
+public interface ProductTypeMapper {
+
+    ProductTypeEntity dtoToEntity(ProductTypeDto dto);
+
+    ProductTypeDto entityToDto(ProductTypeEntity productTypeEntity);
+
+}
