+++ menu-microservice-product/src/main/java/br/com/menu/microservice/product/repository/ProductRepository.java
@@ -1,0 +1,10 @@
+package br.com.menu.microservice.product.repository;
+
+import br.com.menu.microservice.product.model.ProductEntity;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ProductRepository extends CrudRepository<ProductEntity, Long> {
+
+}
