+++ menu-microservice-product/src/main/resources/application.properties
@@ -1,0 +1,9 @@
+spring.datasource.url=jdbc:mysql://mysql:3306/test
+spring.datasource.username=root
+spring.datasource.password=hellokoding
+
+spring.jpa.hibernate.ddl-auto=create-drop
+spring.jpa.database-platform=org.hibernate.dialect.MySQL5Dialect
+spring.jpa.generate-ddl=true
+spring.jpa.show-sql=true
+product.type.rest.url=http://menu-microservice-product-type:8080/menu/productType
